ARM64: t210: jetson: Remove unused rt5659 codec
An rt5659 codec was added to the jetson-cv base dts file apart of some
bring-up (see bug 200118114). However, this codec is no longer used and
is causing the following error message '0-001b: device with id register
0 is not rt5659' to be seen when booting the Jetson TX1 board.

For the Jetson TX1 board the population of the codec is handled by the
plug-in manager when the supermodule is detected and for the Jetson
Erista, which has an on-board rt5640 codec, this is defined in its dts
base file. Therefore, remove this unused codec.

Bug 1926668

Change-Id: Ia774ebeab0d82bacccd919f64c7a01373b5e82d4
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: http://git-master/r/1493856
(cherry picked from commit 2e737b9d044a6b2b9067bb257c90d236ca199600)
Reviewed-on: http://git-master/r/1504927
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Tested-by: Bibek Basu <bbasu@nvidia.com>